requirements.txt
astropy==6.1.4
pandas==2.2.3
numpy==2.1.2
lmfit==1.3.2
xlsxwriter>=3.1.0
matplotlib>=3.5.0
scipy>=1.9.0
GitPython>=3.1.0
pytest
python-dotenv
pytest-cov>=4.0.0
coveralls==2.2.0
barycorrpy>=0.3.0
photutils>=1.5.0
watchdog==2.1.7
astroscrappy==1.2.0
psycopg2-binary>=2.9.0
tqdm>=4.64.0
pyyaml>=6.0
seaborn>=0.11.0
redis>=4.0.0
# The following is required by polly.
pyfonts>=0.0.3

# Documentation
Sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0
nbsphinx>=0.8.0
pandoc>=2.0.0
ipython>=8.0.0
pygments>=2.10.0

# Framework
# --TODO-- tracking master for now. Need to settle on a release in the future
git+https://github.com/Keck-DataReductionPipelines/KeckDRPFramework.git@master

# The following needs to be installed under the KPF-Pipeline directory (so the path is KPF-Pipeline/polly).
# If this is not possible, then the relative path in KPF-Pipeline/cronjobs/run_analysis_for_masters.py
#
# git+https://github.com/pemberj/polly.git@main
